backend/python/summary-service/pyproject.toml
[project]
name = "summary-service"
version = "0.1.0"
description = "Meeting Summarization Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "redis>=5.0.1",
    "openai>=1.10.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "llama-cpp-python>=0.2.23",
    "huggingface-hub>=0.20.0",
]

[tool.setuptools]
py-modules = ["main", "cc_pb2", "cc_pb2_grpc"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
